+++ prototype/preload.js
@@ -43,6 +43,7 @@ contextBridge.exposeInMainWorld('chromux', {
   revealPath: (p) => ipcRenderer.send('reveal-path', { p }),
   setWindowButtonPosition: (position) => ipcRenderer.send('set-window-button-position', position),
   getEnv: () => ipcRenderer.invoke('get-env'),
+  restartWithDevMode: (opts) => ipcRenderer.invoke('restart-with-dev-mode', opts || {}),
   checkUpdates: (opts) => ipcRenderer.invoke('check-updates', opts || {}),
   openUpdateRelease: (opts) => ipcRenderer.invoke('open-update-release', opts || {}),
   openSecurityResource: (resource) => ipcRenderer.invoke('open-security-resource', resource),